+++ frontend/src/routes/auth/index.tsx
@@ -2,12 +2,12 @@ import api from "@sendme/api";
 import { SlSpinner } from "@shoelace-style/shoelace/dist/react";
 import { createFileRoute } from "@tanstack/react-router";
 import styled from "styled-components";
-import RequestPinForm from "./-components/Auth";
+import AuthForm from "./-components/Auth";
 
 export const Route = createFileRoute("/auth/")({
 	component: RouteComponent,
 });
 
 function RouteComponent() {
-	return <RequestPinForm />;
+	return <AuthForm />;
 }
